ml-backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
ultralytics>=8.1.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0
python-multipart>=0.0.6
google-generativeai>=0.4.0
python-dotenv>=1.0.0
